Moved the code managing the graph to a new class. Deactivated the code when double clicking a graph item (needs rework on the SimpleGUI side).

+++ core/Graph.cpp
@@ -19,7 +19,7 @@ void Graph::executeCallback(CallbackReason reason)
 {
 	if(m_updateCallback)
 	{
-		auto val = static_cast<uint16_t>(reason);
+		auto val = static_cast<uint8_t>(reason);
 		m_updateCallback(val);
 	}
 }
